ci(script): update where build:version is being called
