+++ scripts/dependencyTagVersioning.mjs
@@ -6,10 +6,12 @@
 import { literal, unsafeStatic } from 'lit/static-html.js';
 
 export class AuroDependencyVersioning {
+
   /**
-   * Generates a unique string to be used for child auro element naming
-   * private
-   * @param {string} - Defines the first part of the unique element name.
+   * Generates a unique string to be used for child auro element naming.
+   * @private
+   * @param {string} baseName - Defines the first part of the unique element name.
+   * @param {string} version - Version of the component that will be appended to the baseName.
    * @returns {string} - Unique string to be used for naming.
    */
   generateElementName(baseName, version) {
@@ -21,6 +23,12 @@ export class AuroDependencyVersioning {
     return result;
   }
 
+  /**
+   * Generates a unique string to be used for child auro element naming.
+   * @param {string} baseName - Defines the first part of the unique element name.
+   * @param {string} version - Version of the component that will be appended to the baseName.
+   * @returns {string} - Unique string to be used for naming.
+   */
   generateTag(baseName, version, tagClass) {
     const elementName = this.generateElementName(baseName, version);
     const tag = literal`${unsafeStatic(elementName)}`;
